FIX: respect prioritize_username_in_ux site setting
We want to sort the results based on the information available and displayed in the UI.

The prioritize_username_in_ux site setting changes the way we show the username and names, so we need to account for its value when sorting when the cakedate is the same for several users.

The 'tie_breaker' sort is quite a mouthful when we don't prioritize username, but it's ensuring we lower the name (because postgresql order by is case sensitive), then it trims the value and coalesces it to 'username_lower' if the user hasn't provided a name yet.